Refactors numpy integration for improved performance
This commit refactors the numpy integration to directly use the numpy C API,
avoiding unnecessary data copying and improving performance.

It introduces a `PyCArray` struct for safe interaction with numpy arrays
and uses `ecovec` to efficiently transfer data to and from `EcoVec`s.

This change also removes the previous `utils` module,
as its functionality is now integrated into the main module and `ecovec`.